method to put data on pdc
